pyproject.toml
[project]
name = "warporacer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pyyaml>=6.0.3",
    "scikit-image>=0.26.0",
    "scipy>=1.17.1",
    "typer>=0.24.1",
    "warp-lang>=1.12.1",
]
